+++ src/data/config.ts
@@ -13,3 +13,9 @@ export const PROFESSOR_QUERY = {
 
 export const SCHOOL_ID = '1273';
 export const RMP_GRAPHQL_URL = 'https://www.ratemyprofessors.com/graphql';
+
+export const neededOrigins = [
+  'https://utdallas.collegescheduler.com/terms/*/courses/*',
+  'https://www.ratemyprofessors.com/',
+  'https://trends.utdnebula.com/',
+];
